backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
numpy==2.2.2
Pillow==11.3.0
pillow-heif==0.22.0
pydantic==2.11.7
